+++ NEWS.md
@@ -1,6 +1,10 @@
 ## 0.5.2 to 0.5.3 (unreleased)
 
  * Events: Subqueries can be extended by the configuration, see the example configuration.
+ * Events: Subqueries can need the parameter '%s' more or less than one time in the SQL query, for example:
+   ```
+   ("source.ip" = '%s' OR "destination.ip" = '%s')
+   ```
 
 ## 0.5.1 to 0.5.2